:sparkles: feat(e05-t05) move editor to components & update article
